# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Q-Structurize",
    "description": "PDF parsing and structured text extraction API",
    "version": "1.0.0"
  },
  "paths": {
    "/parse/file": {
      "post": {
        "summary": "Parse Pdf File",
        "description": "Parse PDF file and return structured text output.\n\n- **file**: PDF file to parse\n- **max_tokens_per_chunk**: Maximum tokens per chunk (reserved for future use)\n- **optimize_pdf**: Whether to optimize PDF for better text extraction (default: false)",
        "operationId": "parse_pdf_file_parse_file_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_parse_pdf_file_parse_file_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "summary": "Root",
        "description": "Health check endpoint.",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_parse_pdf_file_parse_file_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "max_tokens_per_chunk": {
            "type": "integer",
            "title": "Max Tokens Per Chunk",
            "default": 512
          },
          "optimize_pdf": {
            "type": "boolean",
            "title": "Optimize Pdf",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_parse_pdf_file_parse_file_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ParseResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "data": {
            "type": "object",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "message",
          "status",
          "data"
        ],
        "title": "ParseResponse",
        "description": "Response model for PDF parsing endpoint."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}